Make all extractor class inherit the same ABC.
