+++ data_extractor/abc.py
@@ -4,8 +4,8 @@ Abstract Base Classes.
 # Standard Library
 import warnings
 
-from abc import ABC, abstractmethod
-from typing import Any
+from abc import abstractmethod
+from typing import Any, Dict, List, Tuple
 
 
 class __Sentinel:
@@ -16,7 +16,22 @@ class __Sentinel:
 sentinel = __Sentinel()
 
 
-class AbstractExtractor(ABC):
+class ComplexExtractorMeta(type):
+    """
+    Complex Extractor Meta Class.
+    """
+
+    def __init__(cls, name: str, bases: Tuple[type], attr_dict: Dict[str, Any]):
+        super().__init__(name, bases, attr_dict)
+        field_names: List[str] = []
+        for key, attr in attr_dict.items():
+            if isinstance(type(attr), ComplexExtractorMeta):
+                field_names.append(key)
+
+        cls._field_names = field_names
+
+
+class AbstractExtractor(metaclass=ComplexExtractorMeta):
     def __init__(self, expr: str):
         self.expr = expr
 
@@ -30,6 +45,8 @@ class AbstractExtractor(ABC):
         """
         raise NotImplementedError
 
+
+class ExtractFirstMixin(AbstractExtractor):
     def extract_first(self, element: Any, default: Any = sentinel) -> Any:
         """
         Extract the first data or subelement from `extract` method call result.
@@ -52,4 +69,4 @@ class AbstractExtractor(ABC):
         return rv[0]
 
 
-__all__ = ("AbstractExtractor", "sentinel")
+__all__ = ("AbstractExtractor", "ComplexExtractorMeta", "ExtractFirstMixin", "sentinel")
