+++ tests/test_exceptions.py
@@ -1,0 +1,49 @@
+# Standard Library
+import textwrap
+
+# Third Party Library
+import pytest
+
+# First Party Library
+from data_extractor.exceptions import ExtractError
+from data_extractor.item import Field, Item
+from data_extractor.json import JSONExtractor
+
+
+def test_exception_trace(json0):
+    data = json0
+
+    class User(Item):
+        uid = Field(JSONExtractor("id"))
+        name = Field(JSONExtractor("name"))
+        gender = Field(JSONExtractor("gender"))
+
+    class UserResponse(Item):
+        start = Field(JSONExtractor("start"), default=0)
+        size = Field(JSONExtractor("size"))
+        total = Field(JSONExtractor("total"))
+        data = User(JSONExtractor("users[*]"), is_many=True)
+
+    extractor = UserResponse(JSONExtractor("data"))
+    with pytest.raises(ExtractError) as catch:
+        extractor.extract(data)
+
+    exc = catch.value
+    assert len(exc.extractors) == 3
+    assert exc.extractors[0] is User.gender
+    assert exc.extractors[1] is UserResponse.data
+    assert exc.extractors[2] is extractor
+    assert exc.element == {"id": 3, "name": "Janine Gross"}
+
+    assert (
+        exc._trace_repr
+        == textwrap.dedent(
+            """
+            ExtractError(Field(JSONExtractor('gender'), default=sentinel, is_many=False), element={'id': 3, 'name': 'Janine Gross'})
+            |-UserResponse(JSONExtractor('data'), default=sentinel, is_many=False)
+              |-User(JSONExtractor('users[*]'), default=sentinel, is_many=True)
+                |-Field(JSONExtractor('gender'), default=sentinel, is_many=False)
+                  |-{'id': 3, 'name': 'Janine Gross'}
+            """
+        ).strip()
+    )
